Moved supporting functions for filtering to _filtering file
