+++ cmd/rpcdaemon/commands/trace_api.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/ledgerwatch/turbo-geth/cmd/rpcdaemon/cli"
 	"github.com/ledgerwatch/turbo-geth/common"
-	"github.com/ledgerwatch/turbo-geth/common/dbutils"
 	"github.com/ledgerwatch/turbo-geth/common/hexutil"
 	"github.com/ledgerwatch/turbo-geth/core/rawdb"
 	"github.com/ledgerwatch/turbo-geth/core/types"
@@ -56,61 +55,6 @@ func NewTraceAPI(db ethdb.KV, dbReader ethdb.Getter, cfg *cli.Flags) *TraceAPIIm
 		maxTraces: cfg.MaxTraces,
 		traceType: cfg.TraceType,
 	}
-}
-
-func retrieveHistory(tx ethdb.Tx, addr *common.Address, fromBlock uint64, toBlock uint64) ([]uint64, error) {
-	addrBytes := addr.Bytes()
-	ca := tx.Cursor(dbutils.AccountsHistoryBucket).Prefix(addrBytes)
-	var blockNumbers []uint64
-
-	for k, v, err := ca.First(); k != nil; k, v, err = ca.Next() {
-		if err != nil {
-			return nil, err
-		}
-
-		numbers, _, err := dbutils.WrapHistoryIndex(v).Decode()
-		if err != nil {
-			return nil, err
-		}
-
-		blockNumbers = append(blockNumbers, numbers...)
-	}
-
-	// cleanup for invalid blocks
-	start := -1
-	end := -1
-	for i, b := range blockNumbers {
-		if b >= fromBlock && b <= toBlock && start == -1 {
-			start = i
-			continue
-		}
-
-		if b > toBlock {
-			end = i
-			break
-		}
-	}
-
-	if start == -1 {
-		return []uint64{}, nil
-	}
-
-	if end == -1 {
-		return blockNumbers[start:], nil
-	}
-	// Remove dublicates
-	return blockNumbers[start:end], nil
-}
-
-func isAddressInFilter(addr *common.Address, filter []*common.Address) bool {
-	if filter == nil {
-		return true
-	}
-	i := sort.Search(len(filter), func(i int) bool {
-		return bytes.Equal(filter[i].Bytes(), addr.Bytes())
-	})
-
-	return i != len(filter)
 }
 
 // Filter Implements trace_filter
